Avoid and clear as much taint as possible to stop Action Blocked errors
This particularly tries to avoid tainting ItemButtons and nilling
anything in MerchantFrame tainted by calling the OnClick handlers
after the handlers are confirmed or the merchant is closed.

It also limits the use of UpdateBuyback() to only when absolutely
necessary to correct the widget's state and clears tainted values
on merchant close.