Fix sorting in reverse
Useful programming lesson: Inverting the result of individual
comparisons is not the same as inverting the result of the sort.
It breaks when sorting a column with duplicate values, because
equal values are incorrectly reported as if the RHS is lesser.

Instead, flip the left and right side in comparisons when reverse
sorting.